Fix wheel creation with Secure Control Plane
The laldoc build is broken with the _Secure Control Plane_ since we
switched to using `pyproject.toml` rather than `setup.py`.

While in the _Secure Control Plane_, there is no access to `pypi`, and
we have to use our thirdparties package to be able to build the wheels.

Using `python -m build wheel` was not possible because of the options
to pass to `pip` (even using `-C"--no-cache-dir"` would not work, so
the `Makefile` of `laldoc` has been updated to rather use `pip wheel`.

Ref eng/libadalang/sphinxcontrib-adadomain#6+